requirements.txt
fastapi
Pillow
numpy
uvicorn
python-multipart